+++ Sources/StreamChat/Database/DTOs/AttachmentDTO.swift
@@ -35,7 +35,7 @@ class AttachmentDTO: NSManagedObject {
     /// An attachment local url.
     @NSManaged var localURL: URL?
     /// An attachment raw `Data`.
-    @NSManaged var data: Data?
+    @NSManaged var data: Data
     
     // MARK: - Relationships
     
@@ -125,7 +125,7 @@ extension NSManagedObjectContext: AttachmentDatabaseSession {
         dto.localURL = attachment.localFileURL
         dto.localState = attachment.localFileURL == nil ? .uploaded : .pendingUpload
 
-        dto.data = try JSONEncoder.stream.encode(attachment.payload?.asAnyEncodable)
+        dto.data = try JSONEncoder.stream.encode(attachment.payload.asAnyEncodable)
         dto.channel = channelDTO
         dto.message = messageDTO
         
@@ -154,70 +154,17 @@ private extension AttachmentDTO {
             return nil
         }
     }
-
-    func asModel<T: Decodable>(payloadType: T.Type = T.self) -> _ChatMessageAttachment<T>? {
-        guard
-            let payload = payload(ofType: payloadType)
-        else { return nil }
-
-        return .init(
-            id: attachmentID,
-            type: attachmentType,
-            payload: payload,
-            uploadingState: uploadingState
-        )
-    }
-
-    /// Helper decoding method that logs error only if object exists.
-    /// Returns `nil` if `Data` for decoding is `nil`.
-    func payload<T: Decodable>(ofType type: T.Type = T.self) -> T? {
-        guard let data = data else { return nil }
-
-        do {
-            return try JSONDecoder.default.decode(type, from: data)
-        } catch {
-            log.error(
-                "Failed to decode attachment of type:\(type) with hash: <\(id)>, "
-                    + "falling back to ChatMessageCustomAttachment."
-                    + "Error: \(error)"
-            )
-            return nil
-        }
-    }
 }
 
 extension AttachmentDTO {
     /// Snapshots the current state of `AttachmentDTO` and returns an immutable model object from it.
-    func asAnyModel() -> AnyChatMessageAttachment? {
-        let attachment: AnyChatMessageAttachment?
-
-        switch attachmentType {
-        case .image:
-            attachment = asModel(payloadType: ImageAttachmentPayload.self)?.asAnyAttachment
-        case .file:
-            attachment = asModel(payloadType: FileAttachmentPayload.self)?.asAnyAttachment
-        case .video:
-            attachment = asModel(payloadType: VideoAttachmentPayload.self)?.asAnyAttachment
-        case .giphy:
-            attachment = asModel(payloadType: GiphyAttachmentPayload.self)?.asAnyAttachment
-        case .linkPreview:
-            attachment = asModel(payloadType: LinkAttachmentPayload.self)?.asAnyAttachment
-        default:
-            attachment = data.map {
-                .init(
-                    id: attachmentID,
-                    type: attachmentType,
-                    payload: $0 as Any,
-                    uploadingState: uploadingState
-                )
-            }
-        }
-
-        if attachment == nil {
-            log.error("Failed to decode attachment of type: \(attachmentType)")
-        }
-
-        return attachment
+    func asAnyModel() -> AnyChatMessageAttachment {
+        .init(
+            id: attachmentID,
+            type: attachmentType,
+            payload: data,
+            uploadingState: uploadingState
+        )
     }
     
     /// Snapshots the current state of `AttachmentDTO` and returns its representation for used in API calls.
@@ -225,7 +172,6 @@ extension AttachmentDTO {
     /// That is why `RawJSON` object is used for sending it to backend because SDK doesn't know the structure of custom attachment.
     func asRequestPayload() -> MessageAttachmentPayload? {
         guard
-            let data = data,
             let payload = try? JSONDecoder.default.decode(RawJSON.self, from: data)
         else {
             log.error("Internal error. Unable to decode attachment `data` for sending to backend.")
@@ -236,34 +182,34 @@ extension AttachmentDTO {
     }
 
     func update(uploadedFileURL: URL) {
-        switch attachmentType {
-        case .image:
-            guard var image: ImageAttachmentPayload = payload() else {
-                log.assertionFailure(
-                    "Image payload must be decoded to provide the `imageURL` before sending"
-                )
-                return
-            }
-            image.imageURL = uploadedFileURL
-            data = try? JSONEncoder.stream.encode(image)
-        case .video:
-            guard var video: VideoAttachmentPayload = payload() else {
-                log.assertionFailure(
-                    "Video payload must be decoded to provide the `videoURL` before sending"
-                )
-                return
-            }
-            video.videoURL = uploadedFileURL
-            data = try? JSONEncoder.stream.encode(video)
-        default:
-            guard var file: FileAttachmentPayload = payload() else {
-                log.assertionFailure(
-                    "File payload must be decoded to provide the `assetURL` before sending"
-                )
-                return
-            }
-            file.assetURL = uploadedFileURL
-            data = try? JSONEncoder.stream.encode(file)
+        let attachment = asAnyModel()
+        let updatedPayload: AnyEncodable
+        
+        if let image = attachment.attachment(payloadType: ImageAttachmentPayload.self) {
+            var payload = image.payload
+            payload.imageURL = uploadedFileURL
+            updatedPayload = payload.asAnyEncodable
+        } else if let video = attachment.attachment(payloadType: VideoAttachmentPayload.self) {
+            var payload = video.payload
+            payload.videoURL = uploadedFileURL
+            updatedPayload = payload.asAnyEncodable
+        } else if let file = attachment.attachment(payloadType: FileAttachmentPayload.self) {
+            var payload = file.payload
+            payload.assetURL = uploadedFileURL
+            updatedPayload = payload.asAnyEncodable
+        } else {
+            log.assertionFailure(
+                "Attachment of type \(attachment.type) is not supposed to be updated with uploaded file URL."
+            )
+            return
+        }
+        
+        do {
+            data = try JSONEncoder.stream.encode(updatedPayload)
+        } catch {
+            log.assertionFailure(
+                "Failed to encode updated payload for attachment with id \(attachmentID) after uploading."
+            )
         }
     }
 }