+++ Sources/StreamChat/Models/Attachments/ChatMessageFileAttachment.swift
@@ -22,6 +22,17 @@ public struct FileAttachmentPayload: AttachmentPayload {
     public internal(set) var assetURL: URL
     /// The file itself.
     public let file: AttachmentFile
+    /// An extra data.
+    let extraData: [String: RawJSON]?
+    
+    /// Decodes extra data as an instance of the given type.
+    /// - Parameter ofType: The type an extra data should be decoded as.
+    /// - Returns: Extra data of the given type or `nil` if decoding fails.
+    public func extraData<T: Decodable>(ofType: T.Type = T.self) -> T? {
+        extraData
+            .flatMap { try? JSONEncoder.stream.encode($0) }
+            .flatMap { try? JSONDecoder.stream.decode(T.self, from: $0) }
+    }
 }
 
 extension FileAttachmentPayload: Equatable {}
@@ -30,11 +41,12 @@ extension FileAttachmentPayload: Equatable {}
 
 extension FileAttachmentPayload: Encodable {
     public func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: AttachmentCodingKeys.self)
-
-        try container.encode(title, forKey: .title)
-        try container.encode(assetURL, forKey: .assetURL)
-        try file.encode(to: encoder)
+        var values = extraData ?? [:]
+        values[AttachmentCodingKeys.title.rawValue] = title.map { .string($0) }
+        values[AttachmentCodingKeys.assetURL.rawValue] = .string(assetURL.absoluteString)
+        values[AttachmentFile.CodingKeys.size.rawValue] = .integer(Int(file.size))
+        values[AttachmentFile.CodingKeys.mimeType.rawValue] = file.mimeType.map { .string($0) }
+        try values.encode(to: encoder)
     }
 }
 
@@ -43,19 +55,12 @@ extension FileAttachmentPayload: Encodable {
 extension FileAttachmentPayload: Decodable {
     public init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: AttachmentCodingKeys.self)
-
-        guard
-            let assetURL = try container
-            .decodeIfPresent(String.self, forKey: .assetURL)?
-            .attachmentFixedURL
-        else {
-            throw ClientError.AttachmentDecoding("File attachment must contain `assetURL`")
-        }
-
+        
         self.init(
             title: try container.decodeIfPresent(String.self, forKey: .title),
-            assetURL: assetURL,
-            file: try AttachmentFile(from: decoder)
+            assetURL: try container.decode(URL.self, forKey: .assetURL),
+            file: try AttachmentFile(from: decoder),
+            extraData: try Self.decodeExtraData(from: decoder)
         )
     }
 }