+++ Sources/StreamChatTestTools/Models/Attachments/AnyAttachmentPayload_Mock.swift
@@ -42,11 +42,14 @@ public extension AnyAttachmentPayload {
 
 extension AnyAttachmentPayload: Equatable {
     public static func == (lhs: Self, rhs: Self) -> Bool {
-        let lhsData = try? JSONEncoder.default.encode(lhs.payload?.asAnyEncodable)
-        let rhsData = try? JSONEncoder.default.encode(rhs.payload?.asAnyEncodable)
+        let lhsData = try! JSONEncoder.default.encode(lhs.payload.asAnyEncodable)
+        let lhsJSON = try! JSONDecoder.default.decode(RawJSON.self, from: lhsData)
+
+        let rhsData = try! JSONEncoder.default.encode(rhs.payload.asAnyEncodable)
+        let rhsJSON = try! JSONDecoder.default.decode(RawJSON.self, from: rhsData)
 
         return lhs.type == rhs.type &&
             lhs.localFileURL == rhs.localFileURL &&
-            lhsData == rhsData
+            lhsJSON == rhsJSON
     }
 }